Unwrap history on Result objects
This change adds a more overall consistency to the execution flow. It also means that if the end result of any callable is a Result object, it will concatenated with `results` which ultimately end up in the find Result object's `history`. This is particularly important if some callable builds an action dynamically but with the intention of obeying Signals.

This also changes `after_step` to work only on the output. Its execution happens before flow control so returning a new object (like a Signal or Result) in place of the original output still gives you full control over messing with flow control at this stage should you need to.