Replace `default_context` with `kwargs_overrides`
This was added to sort of mimic `input` in Trailblazer macros. It allows keyword arguments to have values in an action that are not from the context. This allows certain values to be passed in without actually modifying the context. This is useful for providing values that won't come from the context or shadowing values you do not want merged into the context in each leg of the result.